Allow SPI mode to be configured for SPI displays.
NVS variable "display_config" receives new option "spimode".

spimode is an integer value between 0 and 3 as used in struct spi_device_interface_config_t.
According to ESP IDF documentation: SPI mode = "representing a pair of (CPOL, CPHA) configuration".

If spimode is not provided in NVS variable "display_config", it is set to the default value of 0.